requirements.txt
wheel==0.30.0
certifi==2017.11.5
chardet==3.0.4
dateutils==0.6.6
idna==2.6
oauthlib==2.0.6
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
